Separate settings discovery from `SettingsResolver`
