+++ crates/workspace/src/lib.rs
@@ -1,3 +1,4 @@
+pub mod discovery;
 pub mod resolve;
 pub mod settings;
 pub mod toml;
